backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
python-multipart
bcrypt
